backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic[email]==2.5.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
Pillow==10.1.0
python-multipart==0.0.22
ruff==0.1.6
black==23.11.0
pre-commit==3.5.0
httpx==0.27.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
